backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
langchain==0.1.16
langchain-community==0.0.38
chromadb==0.4.24
pypdf==4.1.0
pdfplumber==0.11.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
prisma==0.11.0
asyncpg==0.29.0
bcrypt==4.1.2
passlib==1.7.4
